requirements.txt
fastapi
regex
python-jose
redis
mysql-connector-python
tabulate
pandas
numpy
python-dotenv
psycopg2